@@ -1,5 +1,4 @@
-import { SignedPublicPreKeyType, DeviceType, PreKeyType } from "@privacyresearch/libsignal-protocol-typescript";
-
+import { SignedPublicPreKeyType, DeviceType, PreKeyType } from '@privacyresearch/libsignal-protocol-typescript'
 
 export interface PublicDirectoryEntry {
     identityPubKey: ArrayBuffer
@@ -14,20 +13,20 @@ interface FullDirectoryEntry {
     oneTimePreKeys: PreKeyType[]
 }
 
-export class SignalDirectory{
-    private _data: {[address: string]: FullDirectoryEntry} = {}
+export class SignalDirectory {
+    private _data: { [address: string]: FullDirectoryEntry } = {}
 
-    storeKeyBundle(address: string, bundle: FullDirectoryEntry) : void {
+    storeKeyBundle(address: string, bundle: FullDirectoryEntry): void {
         this._data[address] = bundle
     }
 
     addOneTimePreKeys(address: string, keys: PreKeyType[]): void {
-        this._data[address].oneTimePreKeys.unshift(...keys) 
+        this._data[address].oneTimePreKeys.unshift(...keys)
     }
 
     getPreKeyBundle(address: string): DeviceType | undefined {
         const bundle = this._data[address]
-        if(!bundle) {
+        if (!bundle) {
             return undefined
         }
         const oneTimePreKey = bundle.oneTimePreKeys.pop()
